requirements.txt
matplotlib==3.7.1
matplotlib==3.3.4
numpy==1.20.1
numpy==1.23.0
numpy==1.21.6
opencv_python==4.7.0.72
opencv_python==4.6.0.66
pandas==1.2.4
Pillow==8.2.0
Pillow==9.5.0
Pillow==9.3.0
Pillow==10.0.0
Requests==2.31.0
scikit_learn==0.24.1
tqdm==4.59.0
tqdm==4.65.0
yolov5==7.0.9
